+++ packages/components/src/controllers/useForm.ts
@@ -7,6 +7,8 @@ import type { IconName } from "../components/icon/interfaces";
 import { Status } from "../components/interfaces";
 import { InputComponent, isSupportedType, syncInputDelegate } from "../components/input/common/input";
 import { isCalciteFocusable, SetFocusable } from "../utils/dom";
+import { logger } from "../utils/logger";
+import { validate } from "./useForm/validation";
 
 /** Any form <Component> with a `calcite<Component>Input` event needs to be included in this array. */
 export const componentsWithInputEvent = [
@@ -24,7 +26,6 @@ export const componentsWithInputEvent = [
  *
  * Exported for testing purposes.
  *
- * @param componentTag the tag of the component, e.g. "calcite-input"
  * @returns the event name
  */
 export function getClearValidationEventName(componentTag: string): string {
@@ -39,7 +40,7 @@ export type MutableValidityState = Writable<ValidityState>;
  *
  * Allows calling submit/reset methods on the form.
  */
-interface FormOwner extends LitElement {
+export interface FormOwnerComponent extends LitElement {
   /**
    * The ID of the form to associate with the component.
    *
@@ -47,7 +48,7 @@ interface FormOwner extends LitElement {
    *
    * Note that this prop should use the `@property` decorator.
    */
-  form: string;
+  form?: string;
 }
 
 /**
@@ -57,7 +58,7 @@ interface FormOwner extends LitElement {
  */
 export interface FormComponent<T = any>
   extends
-    FormOwner,
+    FormOwnerComponent,
     LitElement,
     SetFocusable,
     // 👇 needed, otherwise types don't come through when using FormComponent | CheckableFormComponent
@@ -200,12 +201,34 @@ interface UseForm {
   active: boolean;
 
   /**
+   * Helper for overriding the initial default value if not finalized by the first `willUpdate` pass.
+   */
+  overrideDefaultValue: (value: any) => void;
+
+  /**
+   * For components that support multiple input types (e.g. "text", "email", etc.), this method allows changing the input type.
+   */
+  overrideInputType: (type: HTMLInputElement["type"]) => void;
+
+  /**
    * Calls `requestSubmit()` on the associated form, if there is one.
    */
   requestSubmit: () => void;
+
+  /**
+   * Sets the custom validity of the component.
+   */
+  setCustomValidity: (message: string) => void;
 }
 
-interface UseFormOptions {
+export interface UseFormOptions {
+  /**
+   * A function that returns the value to be submitted for this component. If not provided, the controller will attempt to determine the value based on the component's `value` property and, if applicable, `checked` property.
+   *
+   * Note: this is mostly intended for components that need to map their value differently
+   */
+  getValue?: () => any;
+
   /**
    * When set, the component will validate as if it were the specified input type (e.g. "email").
    */
@@ -219,15 +242,14 @@ export const useForm = <T extends FormComponent>(
   options: UseFormOptions,
 ): ReturnType<typeof makeGenericController<UseForm, T>> => {
   return makeGenericController<UseForm, T>((component, controller) => {
-    let defaultValueDirty = false;
-    let defaultCheckedDirty = false;
+    let customValidityMessage = "";
     let inputDelegate: HTMLInputElement | undefined;
     let lastAssociatedForm: HTMLFormElement | null = null;
+    let effectiveInputType = options.inputType;
 
-    if (options.inputType) {
-      inputDelegate = document.createElement("input");
-      inputDelegate.type = options.inputType;
+    if (effectiveInputType) {
       // intentionally not appended to the DOM, we just need it for validation
+      inputDelegate = document.createElement("input");
     }
 
     function invalidFormHandler(event: Event): void {
@@ -256,13 +278,10 @@ export const useForm = <T extends FormComponent>(
       }
 
       if (isCheckable(component)) {
-        component.checked = defaultCheckedDirty ? component.defaultChecked : component.checked;
+        component.checked = component.defaultChecked;
       }
 
-      component.value = defaultValueDirty ? component.defaultValue : component.value;
-
-      defaultValueDirty = false;
-      defaultCheckedDirty = false;
+      component.value = component.defaultValue;
     }
 
     component.listen("luminaFormResetCallback", () => {
@@ -280,7 +299,7 @@ export const useForm = <T extends FormComponent>(
     });
 
     function handleInvalidInput(): void {
-      const validationMsg = inputDelegate?.validationMessage || "";
+      const validationMsg = customValidityMessage || inputDelegate?.validationMessage || "";
 
       component.el.dispatchEvent(
         // allows users to set custom validation messages
@@ -324,58 +343,110 @@ export const useForm = <T extends FormComponent>(
     });
 
     controller.onUpdate((changes: PropertyValues<typeof component>) => {
-      if (changes.has("value") && !defaultValueDirty) {
-        defaultValueDirty = true;
+      if (!component.hasUpdated) {
         component.defaultValue = component.value;
+
+        if (isCheckable(component)) {
+          component.defaultChecked = component.checked;
+        }
       }
 
-      if (isCheckable(component) && changes.has("checked") && !defaultCheckedDirty) {
-        defaultCheckedDirty = true;
-        component.defaultChecked = component.checked;
-      }
-
-      if (changes.has("value") || (isCheckable(component) && changes.has("checked"))) {
+      if (changes.has("name") || changes.has("value") || (isCheckable(component) && changes.has("checked"))) {
         component.elementInternals.setFormValue(getFormValue());
       }
 
-      if (inputDelegate) {
-        inputDelegate.value = component.value;
-        syncInternalInput(component, inputDelegate);
-        inputDelegate.checkValidity();
-        component.elementInternals.setValidity(inputDelegate.validity, inputDelegate.validationMessage);
-        if ("validity" in component) {
-          component.validity = component.elementInternals.validity;
-        }
-      }
+      updateValidity();
     });
 
+    function updateValidity(): void {
+      const { elementInternals } = component;
+
+      let validity: ValidityStateFlags = {};
+      let validationMessage = "";
+
+      if (inputDelegate) {
+        inputDelegate.type = effectiveInputType!;
+        syncInternalInput(component, inputDelegate);
+        ({ validity, validationMessage } = validate(inputDelegate, getComponentValue()));
+      }
+
+      if (customValidityMessage) {
+        validity = { ...validity, customError: true };
+        validationMessage = customValidityMessage;
+      }
+
+      elementInternals.setValidity(validity, validationMessage);
+
+      if ("validity" in component) {
+        component.validity = elementInternals.validity;
+      }
+    }
+
+    function getComponentValue(): any {
+      if (options.getValue) {
+        return options.getValue();
+      }
+
+      return component.value;
+    }
+
     function getFormValue(): any {
-      if (Array.isArray(component.value)) {
+      const value = getComponentValue();
+
+      if (Array.isArray(value)) {
         const formData = new FormData();
-        component.value.forEach((value) => formData.append(component.name, value));
+        value.forEach((value) => formData.append(component.name, value));
         return formData;
       }
 
       if (isCheckable(component)) {
         if (component.checked) {
           // matches https://html.spec.whatwg.org/multipage/input.html#dom-input-value-default-on
-          return component.defaultValue === undefined && !defaultValueDirty && component.value === undefined
-            ? "on"
-            : component.value;
+          return value || "on";
         }
 
         return null;
       }
 
-      return component.value;
+      return value;
     }
+
+    let defaultValueOverridden = false;
 
     return {
       get active() {
         return !!component.elementInternals.form;
       },
+      overrideDefaultValue: (value) => {
+        if (import.meta.env.DEV) {
+          if (defaultValueOverridden) {
+            logger.warn("Default value override has already been set.");
+            return;
+          }
+
+          if (!component.hasUpdated) {
+            logger.warn("Overriding before first update is not necessary.");
+            return;
+          }
+        }
+
+        component.defaultValue = value;
+        defaultValueOverridden = true;
+      },
+      overrideInputType: (type) => {
+        if (import.meta.env.DEV && !inputDelegate) {
+          throw new Error("Cannot override input type because no input delegate is configured.");
+        }
+
+        effectiveInputType = type;
+        updateValidity();
+      },
       requestSubmit: () => {
         component.elementInternals.form?.requestSubmit();
+      },
+      setCustomValidity: (message) => {
+        customValidityMessage = message;
+        updateValidity();
       },
     };
   });