+++ packages/components/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.1.0-next.24](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.23...@esri/calcite-components@5.1.0-next.24) (2026-03-27)
+
+**Note:** Version bump only for package @esri/calcite-components
+
+## [5.1.0-next.23](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.22...@esri/calcite-components@5.1.0-next.23) (2026-03-26)
+
+### Features
+
+- **autocomplete-item:** add selected property ([#14137](https://github.com/Esri/calcite-design-system/issues/14137)), closes [#14058](https://github.com/Esri/calcite-design-system/issues/14058)
+
+## [5.1.0-next.22](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.21...@esri/calcite-components@5.1.0-next.22) (2026-03-25)
+
+### Bug Fixes
+
+- **meter:** update border color ([#14167](https://github.com/Esri/calcite-design-system/issues/14167)), closes [#14166](https://github.com/Esri/calcite-design-system/issues/14166)
+
+## [5.1.0-next.21](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.20...@esri/calcite-components@5.1.0-next.21) (2026-03-25)
+
+### Bug Fixes
+
+- **switch:** remove double focus ring in certain browsers ([#14156](https://github.com/Esri/calcite-design-system/issues/14156)), closes [#10824](https://github.com/Esri/calcite-design-system/issues/10824) [#10824](https://github.com/Esri/calcite-design-system/issues/10824)
+
+## [5.1.0-next.20](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.19...@esri/calcite-components@5.1.0-next.20) (2026-03-23)
+
+### Features
+
+- **sort-handle:** hide reorder group title when no sibling groups; disable boundary items instead of hiding ([#14105](https://github.com/Esri/calcite-design-system/issues/14105)), closes [#11451](https://github.com/Esri/calcite-design-system/issues/11451)
+
+## [5.1.0-next.19](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.18...@esri/calcite-components@5.1.0-next.19) (2026-03-23)
+
+### Bug Fixes
+
+- **dropdown:** focus the first item when opening instead of the first selected item ([#14131](https://github.com/Esri/calcite-design-system/issues/14131)), closes [#14128](https://github.com/Esri/calcite-design-system/issues/14128)
+
+## [5.1.0-next.18](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.17...@esri/calcite-components@5.1.0-next.18) (2026-03-23)
+
+**Note:** Version bump only for package @esri/calcite-components
+
+## [5.1.0-next.17](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.16...@esri/calcite-components@5.1.0-next.17) (2026-03-20)
+
+### Features
+
+- **tooltip:** add `pointerDisabled` and `scale` properties ([#14078](https://github.com/Esri/calcite-design-system/issues/14078)), closes [#5334](https://github.com/Esri/calcite-design-system/issues/5334)
+
+## [5.1.0-next.16](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.15...@esri/calcite-components@5.1.0-next.16) (2026-03-20)
+
+### Features
+
+- **dropdown, split-button, sort-handle:** add additional placement options ([#14132](https://github.com/Esri/calcite-design-system/issues/14132)), closes [#4124](https://github.com/Esri/calcite-design-system/issues/4124)
+
+## [5.1.0-next.15](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.14...@esri/calcite-components@5.1.0-next.15) (2026-03-20)
+
+### Features
+
+- add custom Stylelint plugin as private package ([#14107](https://github.com/Esri/calcite-design-system/issues/14107)), closes [#13677](https://github.com/Esri/calcite-design-system/issues/13677)
+
+## [5.1.0-next.14](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.13...@esri/calcite-components@5.1.0-next.14) (2026-03-20)
+
+### Bug Fixes
+
+- **dropdown:** keep focus on trigger with single-component keyboard navigation ([#14079](https://github.com/Esri/calcite-design-system/issues/14079)), closes [#8206](https://github.com/Esri/calcite-design-system/issues/8206) [#13997](https://github.com/Esri/calcite-design-system/issues/13997) [#8206](https://github.com/Esri/calcite-design-system/issues/8206)
+
+## [5.1.0-next.13](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.1.0-next.12...@esri/calcite-components@5.1.0-next.13) (2026-03-19)
+
+### Bug Fixes
+
+- **date-picker:** show adjacent-month days in single-calendar range ([#14111](https://github.com/Esri/calcite-design-system/issues/14111)), closes [#14110](https://github.com/Esri/calcite-design-system/issues/14110)
+- **popover, tooltip:** correctly assign aria properties ([#14098](https://github.com/Esri/calcite-design-system/issues/14098)), closes [#13242](https://github.com/Esri/calcite-design-system/issues/13242) [#13242](https://github.com/Esri/calcite-design-system/issues/13242)
+
+## 5.1.0-next.12 (2026-03-19)
+
+- fix(stepper): fix updating items when component is used within a shadow dom and items are slotted (# ([bc7a74a](https://github.com/Esri/calcite-design-system/commit/bc7a74a)), closes [#14021](https://github.com/Esri/calcite-design-system/issues/14021) [#11737](https://github.com/Esri/calcite-design-system/issues/11737) [#11737](https://github.com/Esri/calcite-design-system/issues/11737)
+- fix(input, input-number): fix handling of input for locales with same group and decimal separators ( ([2dae3be](https://github.com/Esri/calcite-design-system/commit/2dae3be)), closes [#14112](https://github.com/Esri/calcite-design-system/issues/14112) [#14100](https://github.com/Esri/calcite-design-system/issues/14100)
+- refactor(alert, dialog, sheet, tree-item): swap to a unified animation curve (#14038) ([6ba56e0](https://github.com/Esri/calcite-design-system/commit/6ba56e0)), closes [#14038](https://github.com/Esri/calcite-design-system/issues/14038) [#12669](https://github.com/Esri/calcite-design-system/issues/12669)
+- refactor(input): port tests to browser mode (#14104) ([09d8ff9](https://github.com/Esri/calcite-design-system/commit/09d8ff9)), closes [#14104](https://github.com/Esri/calcite-design-system/issues/14104) [#11268](https://github.com/Esri/calcite-design-system/issues/11268)
+- refactor(input): wire up `useForm` (#14073) ([0d37d77](https://github.com/Esri/calcite-design-system/commit/0d37d77)), closes [#14073](https://github.com/Esri/calcite-design-system/issues/14073) [#8126](https://github.com/Esri/calcite-design-system/issues/8126)
+- refactor(text-area): wire up `useForm` (#14074) ([5ba8b04](https://github.com/Esri/calcite-design-system/commit/5ba8b04)), closes [#14074](https://github.com/Esri/calcite-design-system/issues/14074) [#8126](https://github.com/Esri/calcite-design-system/issues/8126)
+- test(list): unskip test (#14102) ([31216a4](https://github.com/Esri/calcite-design-system/commit/31216a4)), closes [#14102](https://github.com/Esri/calcite-design-system/issues/14102) [#11292](https://github.com/Esri/calcite-design-system/issues/11292)
+- docs: deprecate --calcite-stepper-bar-inactive-fill-color and --calcite-stepper-bar-active-fill-colo ([1b3c47d](https://github.com/Esri/calcite-design-system/commit/1b3c47d)), closes [#14096](https://github.com/Esri/calcite-design-system/issues/14096) [#13720](https://github.com/Esri/calcite-design-system/issues/13720)
+
 ## 5.1.0-next.11 (2026-03-16)
 
 - fix(input-text): restore focus when it is the first invalid element on form submit (#14086) ([ac8f0db](https://github.com/Esri/calcite-design-system/commit/ac8f0db)), closes [#14086](https://github.com/Esri/calcite-design-system/issues/14086)